server/requirements.txt
routes
cherrypy
sqlalchemy
six
eventlet
pyjwt
pymysql

